move : move gameData file in default folder
